+++ crates/core/utils/src/lib.rs
@@ -66,34 +66,6 @@ pub fn is_docker_container() -> bool {
     PathBuf::from("/.dockerenv").exists()
 }
 
-// TODO: This doesn't work behind VPN or corporate proxies. Disabling for now
-// until we can figure out a better solution.
-#[cached(time = 60)]
-pub fn is_offline() -> bool {
-    false
-    // use std::time::Duration;
-    // use std::net::{Shutdown, SocketAddr, TcpStream};
-
-    // let addresses = [
-    //     // Cloudflare DNS: https://1.1.1.1/dns/
-    //     SocketAddr::from(([1, 1, 1, 1], 53)),
-    //     SocketAddr::from(([1, 0, 0, 1], 53)),
-    //     // Google DNS: https://developers.google.com/speed/public-dns
-    //     SocketAddr::from(([8, 8, 8, 8], 53)),
-    //     SocketAddr::from(([8, 8, 4, 4], 53)),
-    // ];
-
-    // for address in addresses {
-    //     if let Ok(stream) = TcpStream::connect_timeout(&address, Duration::new(3, 0)) {
-    //         stream.shutdown(Shutdown::Both).unwrap();
-
-    //         return false;
-    //     }
-    // }
-
-    // true
-}
-
 #[inline]
 pub fn is_test_env() -> bool {
     env::var("MOON_TEST").is_ok()