+++ crates/proto/core/src/resolver.rs
@@ -1,8 +1,8 @@
-use crate::color;
+use crate::errors::ProtoError;
+use crate::{color, is_offline};
 use crate::{get_temp_dir, is_version_alias, remove_v_prefix};
 use lenient_semver::Version;
 use log::trace;
-use proto_error::ProtoError;
 use serde::de::DeserializeOwned;
 use sha2::{Digest, Sha256};
 use std::collections::BTreeMap;
@@ -86,6 +86,9 @@ pub trait Resolvable<'tool>: Send + Sync {
     /// Given an initial version, resolve it to a fully qualifed and semantic version
     /// according to the tool's ecosystem.
     async fn resolve_version(&mut self, initial_version: &str) -> Result<String, ProtoError>;
+
+    /// Explicitly set the resolved version.
+    fn set_version(&mut self, version: &str);
 }
 
 pub async fn load_versions_manifest<T, U>(url: U) -> Result<T, ProtoError>
@@ -101,27 +104,37 @@ where
     let temp_file = temp_dir.join(format!("{:x}.json", sha.finalize()));
     let handle_http_error = |e: reqwest::Error| ProtoError::Http(url.to_owned(), e.to_string());
     let handle_io_error = |e: io::Error| ProtoError::Fs(temp_file.to_path_buf(), e.to_string());
+    let offline = is_offline();
 
-    // If the resource has been cached within the last 24 hours, use it
     if temp_file.exists() {
         let metadata = fs::metadata(&temp_file).map_err(handle_io_error)?;
 
-        if let Ok(modified_time) = metadata.modified().or_else(|_| metadata.created()) {
-            let threshold = SystemTime::now() - Duration::from_secs(60 * 60 * 24);
+        // When offline, always read the temp file as we can't download the manifest
+        let read_temp = if offline {
+            true
+            // Otherwise, only read the temp file if its been downloaded in the last 24 hours
+        } else if let Ok(modified_time) = metadata.modified().or_else(|_| metadata.created()) {
+            modified_time > SystemTime::now() - Duration::from_secs(60 * 60 * 24)
+        } else {
+            false
+        };
 
-            if modified_time > threshold {
-                trace!(
-                    target: "proto:resolver",
-                    "Loading versions manifest from locally cached {}",
-                    color::path(&temp_file),
-                );
+        if read_temp {
+            trace!(
+                target: "proto:resolver",
+                "Loading versions manifest from locally cached {}",
+                color::path(&temp_file),
+            );
 
-                let contents = fs::read_to_string(&temp_file).map_err(handle_io_error)?;
+            let contents = fs::read_to_string(&temp_file).map_err(handle_io_error)?;
 
-                return serde_json::from_str(&contents)
-                    .map_err(|e| ProtoError::Fs(temp_file.to_path_buf(), e.to_string()));
-            }
+            return serde_json::from_str(&contents)
+                .map_err(|e| ProtoError::Fs(temp_file.to_path_buf(), e.to_string()));
         }
+    }
+
+    if offline {
+        return Err(ProtoError::InternetConnectionRequired);
     }
 
     // Otherwise, request the resource and cache it
@@ -143,4 +156,8 @@ where
 pub fn parse_version(version: &str) -> Result<Version, ProtoError> {
     Version::parse(version)
         .map_err(|e| ProtoError::VersionParseFailed(version.to_owned(), e.to_string()))
+}
+
+pub fn is_semantic_version(version: &str) -> bool {
+    Version::parse(version).is_ok()
 }