+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Updated `node.version` and `node.<packageManager>.version` to no longer default to a hard-coded
   version. When not defined, will fallback to the binary available on `PATH`.
 
+#### 🚀 Updates
+
+- Added offline (no internet connection) support.
+
 #### 🐞 Fixes
 
 - Fixed an issue where `moon query projects --affected` would hang indefinitely waiting for stdin.